jetson_scripts/requirements.txt
annotated-types==0.7.0
anyio==4.12.0
boto3==1.42.10
botocore==1.42.10
cachetools==6.2.4
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
cryptography==46.0.3
deprecation==2.1.0
fsspec==2025.12.0
h11==0.16.0
h2==4.3.0
hpack==4.1.0
httpcore==1.0.9
httpx==0.28.1
hyperframe==6.1.0
idna==3.11
jmespath==1.0.1
markdown-it-py==4.0.0
mdurl==0.1.2
mmh3==5.2.0
multidict==6.7.0
packaging==25.0
postgrest==2.26.0
propcache==0.4.1
psutil==7.1.3
pycparser==2.23
pydantic==2.12.5
pydantic_core==2.41.5
Pygments==2.19.2
pyiceberg==0.10.0
PyJWT==2.10.1
pyparsing==3.2.5
pyroaring==1.0.3
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
realtime==2.26.0
requests==2.32.5
rich==14.2.0
s3transfer==0.16.0
six==1.17.0
sortedcontainers==2.4.0
storage3==2.26.0
StrEnum==0.4.15
strictyaml==1.7.3
supabase==2.26.0
supabase-auth==2.26.0
supabase-functions==2.26.0
tenacity==9.1.2
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.2
websockets==15.0.1
yarl==1.22.0
